requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.25.0
qwen-vl-utils>=0.0.8

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# 2D barcode (codebook_external baseline)
qrcode>=7.4.0
pylibdmtx>=0.1.10      # DataMatrix encode/decode
pyzbar>=0.1.9          # optional: QR decode fallback (needs zbar installed)

# Config & IO
pyyaml>=6.0
jsonlines>=4.0.0

# LoRA training
peft>=0.12.0
bitsandbytes>=0.41.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Utils
tqdm>=4.66.0
